+++ src/main.tsx
@@ -2,12 +2,15 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { App } from "./App";
 import { I18nProvider } from "./i18n/I18nContext";
+import { SettingsProvider } from "./settings/SettingsContext";
 import "./index.css";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <I18nProvider>
-      <App />
+      <SettingsProvider>
+        <App />
+      </SettingsProvider>
     </I18nProvider>
   </StrictMode>,
 );